update: 重构 convert/attrs.js
